Extract fleet inference refinement into inferred_acquisition_refine module.
Placeholder creation stays in ingest; option-set refinement and per-builtTurn scores lookup move to a focused sibling module with no behavior change.
